0902 ORDER_PRODUCT_VIEW & DTO

+++ src/main/java/com/example/k5_iot_springboot/entity/view/OrderSummaryView.java
@@ -30,7 +30,7 @@ public class OrderSummaryView {
     private String product_name;
     private Integer quantity;
     private Integer price;
-    private Integer total_price;
+    private Long total_price;
 
     @Column(name = "ordered_at")
     private LocalDateTime orderedAt;
